--- src/main/java/app/MainForkYeahApplication.java
@@ -1,6 +1,0 @@
-package app;
-
-import view.LoginPageView;
-
-public class MainForkYeahApplication {
-}
