--- src/main/java/data_access/UserDataAccess.java
@@ -1,7 +1,0 @@
-package data_access;
-
-import entity.User;
-
-public interface UserDataAccess {
-    User findUser(String username);
-}
