--- src/main/java/entity/User.java
@@ -1,24 +1,0 @@
-package entity;
-
-/**
- * The representation of a password-protected user for our program.
- */
-public class User {
-
-    private final String name;
-    private final String password;
-
-    public User(String name, String password) {
-        this.name = name;
-        this.password = password;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-}
