--- src/main/java/interface_adapter/ViewModel.java
@@ -1,63 +1,0 @@
-package interface_adapter;
-
-import java.beans.PropertyChangeListener;
-import java.beans.PropertyChangeSupport;
-
-/**
- * The ViewModel for our CA implementation.
- * This class delegates work to a PropertyChangeSupport object for
- * managing the property change events.
- *
- * @param <T> The type of state object contained in the model.
- */
-public class ViewModel<T> {
-
-    private final String viewName;
-
-    private final PropertyChangeSupport support = new PropertyChangeSupport(this);
-
-    private T state;
-
-    public ViewModel(String viewName) {
-        this.viewName = viewName;
-    }
-
-    public String getViewName() {
-        return this.viewName;
-    }
-
-    public T getState() {
-        return this.state;
-    }
-
-    public void setState(T state) {
-        this.state = state;
-    }
-
-    /**
-     * Fires a property changed event for the state of this ViewModel.
-     */
-    public void firePropertyChanged() {
-        this.support.firePropertyChange("state", null, this.state);
-    }
-
-    /**
-     * Fires a property changed event for the state of this ViewModel, which
-     * allows the user to specify a different propertyName. This can be useful
-     * when a class is listening for multiple kinds of property changes.
-     * <p>For example, the LoggedInView listens for two kinds of property changes;
-     * it can use the property name to distinguish which property has changed.</p>
-     * @param propertyName the label for the property that was changed
-     */
-    public void firePropertyChanged(String propertyName) {
-        this.support.firePropertyChange(propertyName, null, this.state);
-    }
-
-    /**
-     * Adds a PropertyChangeListener to this ViewModel.
-     * @param listener The PropertyChangeListener to be added
-     */
-    public void addPropertyChangeListener(PropertyChangeListener listener) {
-        this.support.addPropertyChangeListener(listener);
-    }
-}
