--- src/main/java/interface_adapter/note/NotePresenter.java
@@ -1,38 +1,0 @@
-package interface_adapter.note;
-
-import use_case.note.NoteOutputBoundary;
-
-/**
- * The presenter for our Note viewing and editing program.
- */
-public class NotePresenter implements NoteOutputBoundary {
-
-    private final NoteViewModel noteViewModel;
-
-    public NotePresenter(NoteViewModel noteViewModel) {
-        this.noteViewModel = noteViewModel;
-    }
-
-    /**
-     * Prepares the success view for the Note related Use Cases.
-     *
-     * @param note the output data
-     */
-    @Override
-    public void prepareSuccessView(String note) {
-        noteViewModel.getState().setNote(note);
-        noteViewModel.getState().setError(null);
-        noteViewModel.firePropertyChanged();
-    }
-
-    /**
-     * Prepares the failure view for the Note related Use Cases.
-     *
-     * @param errorMessage the explanation of the failure
-     */
-    @Override
-    public void prepareFailView(String errorMessage) {
-        noteViewModel.getState().setError(errorMessage);
-        noteViewModel.firePropertyChanged();
-    }
-}
