+++ src/main/java/login/data_access/UserDataAccess.java
@@ -1,0 +1,8 @@
+package login.data_access;
+
+import login.entities.User;
+
+public interface UserDataAccess {
+    User findUser(String username);
+    void saveUser(User user);
+}
