+++ src/main/java/login/entities/User.java
@@ -1,0 +1,19 @@
+package login.entities;
+
+public class User {
+    private final String username;
+    private final String password;
+
+    public User(String username, String password) {
+        this.username = username;
+        this.password = password;
+    }
+
+    public String getUsername() {
+        return username;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+}
