+++ src/main/java/login/use_case/LoginInput.java
@@ -1,0 +1,19 @@
+package login.use_case;
+
+public class LoginInput {
+    private String username;
+    private String password;
+
+    public LoginInput(String username, String password) {
+        this.username = username;
+        this.password = password;
+    }
+
+    public String getUsername() {
+        return username;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+}
