+++ src/main/java/login/use_case/LoginInputBoundary.java
@@ -1,0 +1,5 @@
+package login.use_case;
+
+public interface LoginInputBoundary {
+    void execute(LoginInput input);
+}
