+++ src/main/java/login/use_case/LoginOutputBoundary.java
@@ -1,0 +1,6 @@
+package login.use_case;
+
+public interface LoginOutputBoundary {
+    void prepareSuccessView(String message);
+    void prepareFailView(String message);
+}
