+++ src/main/java/login/use_case/SignupInputBoundary.java
@@ -1,0 +1,5 @@
+package login.use_case;
+
+public interface SignupInputBoundary {
+    void execute(SignupInput input);
+}
