+++ src/main/java/login/use_case/SignupOutputBoundary.java
@@ -1,0 +1,6 @@
+package login.use_case;
+
+public interface SignupOutputBoundary {
+    void prepareSuccessView(String message);
+    void prepareFailView(String message);
+}
