--- src/main/java/use_case/note/DataAccessException.java
@@ -1,10 +1,0 @@
-package use_case.note;
-
-/**
- * Exception thrown when there is an error with accessing data.
- */
-public class DataAccessException extends Exception {
-    public DataAccessException(String string) {
-        super(string);
-    }
-}
