--- src/main/java/use_case/note/LoginInput.java
@@ -1,20 +1,0 @@
-package use_case.note;
-
-public class LoginInput {
-    private String username;
-    private String password;
-
-    public LoginInput(String username, String password) {
-        this.username = username;
-        this.password = password;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-}
-
