--- src/main/java/use_case/note/LoginInputBoundary.java
@@ -1,5 +1,0 @@
-package use_case.note;
-
-public interface LoginInputBoundary {
-    void execute(LoginInput input);
-}
