--- src/main/java/use_case/note/LoginOutputBoundary.java
@@ -1,6 +1,0 @@
-package use_case.note;
-
-public interface LoginOutputBoundary {
-    void prepareSuccessView(String message);
-    void prepareFailView(String message);
-}
