--- src/main/java/use_case/note/SignupInputBoundary.java
@@ -1,5 +1,0 @@
-package use_case.note;
-
-public interface SignupInputBoundary {
-    void execute(SignupInput input);
-}
