--- src/main/java/use_case/note/SignupOutputBoundary.java
@@ -1,6 +1,0 @@
-package use_case.note;
-
-public interface SignupOutputBoundary {
-    void prepareSuccessView(String message);
-    void prepareFailView(String message);
-}
